Conflicts:
	src/client/main.cpp

+++ src/client/main.cpp
@@ -1,4 +1,13 @@
 /*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * This file is part of zmqpp.
+ * Copyright (c) 2011-2015 Contributors as noted in the AUTHORS file.
+ */
+
+/*
  *  Created on: 16 Aug 2011
  *      Author: @benjamg
  */
@@ -33,7 +42,11 @@ int main(int argc, char const* argv[])
 	if( options.show_usage || options.show_help )
 	{
 		show_usage( std::cout, BUILD_CLIENT_NAME );
-		if( options.show_help ) { std::cout << std::endl; show_help( std::cout ); }
+		if( options.show_help )
+		{
+			std::cout << std::endl;
+			show_help( std::cout );
+		}
 
 		return EXIT_FAILURE;
 	}
@@ -225,7 +238,7 @@ int main(int argc, char const* argv[])
 				while( result && (length = strlen( buffer.data() ) - 1) > 0 ) // trim newline from gets
 				{
 					buffer[length] = 0;
-					message.add_raw( buffer.data(), length );
+					message.add_raw( buffer.data(), static_cast<uint64_t>(length) );
 
 					if( options.singlepart ) { break; }
